server/webserver/requirements.txt
fastapi
uvicorn[standard]
jinja2
paho-mqtt
mysql-connector-python
python-dotenv
